split up depcode.py into three files
